# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Rythmx API
  version: 1.0.0
  description: >
    Self-hosted music automation and library management API.
    All endpoints under `/api/v1/` follow a standard envelope:
    success responses include `"status": "ok"` plus payload fields;
    error responses include `"status": "error"` and `"message"`.
    Background-triggering endpoints return HTTP 202 with `"is_alive": bool`.
servers:
  - url: /
    description: Rythmx server (all paths are absolute)

# All endpoints require an API key (X-Api-Key header) except /health
security:
  - ApiKeyAuth: []
...
paths:
...
  /api/v1/settings:
    get:
      tags: [settings]
      summary: Get all application settings
      operationId: getSettings
      responses:
        '200':
          description: Current settings (sensitive keys are masked)
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccess'
                  - $ref: '#/components/schemas/Settings'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/v1/settings/test-lastfm:
    post:
      tags: [settings]
      summary: Test Last.fm connection
      operationId: testLastfm
      responses:
        '200':
          description: Connection test result
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccess'
                  - $ref: '#/components/schemas/ConnectionStatus'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/v1/settings/test-plex:
    post:
      tags: [settings]
      summary: Test Plex connection
      operationId: testPlex
      responses:
        '200':
          description: Connection test result
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccess'
                  - $ref: '#/components/schemas/ConnectionStatus'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/v1/settings/test-soulsync:
    post:
      tags: [settings]
      summary: Test SoulSync connection
      operationId: testSoulsync
      responses:
        '200':
          description: Connection test result
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccess'
                  - $ref: '#/components/schemas/ConnectionStatus'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/v1/settings/test-spotify:
    post:
      tags: [settings]
      summary: Test Spotify API credentials
      operationId: testSpotify
      responses:
        '200':
          description: Connection test result
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccess'
                  - $ref: '#/components/schemas/ConnectionStatus'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/v1/settings/test-fanart:
    post:
      tags: [settings]
      summary: Test Fanart.tv API key
      operationId: testFanart
      responses:
        '200':
          description: Connection test result
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccess'
                  - $ref: '#/components/schemas/ConnectionStatus'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/v1/settings/library-platform:
    post:
      tags: [settings]
      summary: Set the library platform
      operationId: setLibraryPlatform
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [platform]
              properties:
                platform:
                  $ref: '#/components/schemas/LibraryPlatform'
      responses:
        '200':
          $ref: '#/components/responses/Ok'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/v1/settings/clear-history:
    post:
      tags: [settings]
      summary: Clear Cruise Control run history
      operationId: clearHistory
      responses:
        '200':
          $ref: '#/components/responses/Ok'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/v1/settings/reset-db:
    post:
      tags: [settings]
      summary: Reset the entire Rythmx database
      operationId: resetDb
      responses:
        '200':
          $ref: '#/components/responses/Ok'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/v1/settings/clear-image-cache:
    post:
      tags: [settings]
      summary: Clear the image resolution cache
      operationId: clearImageCache
      responses:
        '200':
          $ref: '#/components/responses/Ok'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/v1/settings/api-key:
    get:
      tags: [settings]
      summary: Get the current API key
      operationId: getApiKey
      responses:
        '200':
          description: Current API key
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccess'
                  - type: object
                    required: [api_key]
                    properties:
                      api_key:
                        type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/v1/settings/regenerate-api-key:
    post:
      tags: [settings]
      summary: Regenerate the API key
      description: >
        Generates a new cryptographically random API key and replaces the existing one.
        The caller must update their stored key — the old key is immediately invalidated.
      operationId: regenerateApiKey
      responses:
        '200':
          description: Newly generated API key
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccess'
                  - type: object
                    required: [api_key]
                    properties:
                      api_key:
                        type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/v1/library/status:
    get:
      tags: [settings]
      summary: Get library sync status
      operationId: getLibraryStatus
      responses:
        '200':
          description: Library sync status
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccess'
                  - $ref: '#/components/schemas/LibraryStatus'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/v1/library/sync:
    post:
      tags: [settings]
      summary: Trigger library sync
      operationId: syncLibrary
      responses:
        '202':
          $ref: '#/components/responses/Accepted'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/ServerError'

  # ── Library Browse ────────────────────────────────────────────────────────
...
  /api/v1/library/enrich/status:
    get:
      tags: [library]
      summary: Unified enrichment pipeline status (all workers)
      operationId: getEnrichStatus
      responses:
        '200':
          description: Pipeline status grouped by enrichment source
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccess'
                  - $ref: '#/components/schemas/EnrichmentPipelineStatus'
        '500':
          $ref: '#/components/responses/ServerError'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
      description: >
        Auto-generated API key. View or regenerate in Settings.
        Included as `X-Api-Key: <key>` on every request.
...
  responses:
    Unauthorized:
      description: Missing or invalid API key
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Ok:
      description: Operation succeeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiSuccess'
    Accepted:
      description: Background operation triggered
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ApiSuccess'
              - type: object
                required: [is_alive]
                properties:
                  is_alive:
                    type: boolean
    BadRequest:
      description: Invalid request body or parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
    Conflict:
      description: Operation already in progress or resource conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
  schemas:
...
    LibraryPlatform:
      type: string
      enum: [plex, jellyfin, navidrome]
...
    ApiSuccess:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [ok]
    ApiError:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [error]
        message:
          type: string
        error:
          type: string

    # ── Music Entities ───────────────────────────────────────────────────────
...
    Settings:
      type: object
      required: [library_platform]
      properties:
        lastfm_username:
          type: string
        lastfm_api_key:
          type: string
        plex_url:
          type: string
        plex_token:
          type: string
        soulsync_url:
          type: string
        spotify_client_id:
          type: string
        spotify_client_secret:
          type: string
        library_platform:
          $ref: '#/components/schemas/LibraryPlatform'
    ConnectionStatus:
      type: object
      required: [connected]
      properties:
        connected:
          type: boolean
        message:
          type: string

    # ── Library Enrichment ───────────────────────────────────────────────────
    LibraryStatus:
      type: object
      required: [platform, track_count]
      properties:
        platform:
          $ref: '#/components/schemas/LibraryPlatform'
        track_count:
          type: integer
        last_synced:
          type: string
          format: date-time
        synced:
          type: boolean
        total_albums:
          type: integer
        enriched_albums:
          type: integer
        enrich_pct:
          type: number
          format: float
...
    EnrichmentWorkerStatus:
      type: object
      required: [found, not_found, errors, pending]
      properties:
        found:
          type: integer
        not_found:
          type: integer
        errors:
          type: integer
        pending:
          type: integer
    EnrichmentPipelineStatus:
      type: object
      required: [running]
      properties:
        running:
          type: boolean
        started_at:
          type: [string, 'null']
          format: date-time
          description: ISO 8601 UTC timestamp set when run_full() is called; null when not running
        workers:
          type: object
          description: Per-worker progress keyed by worker name
          properties:
            library:        { $ref: '#/components/schemas/EnrichmentWorkerStatus' }
            itunes_rich:    { $ref: '#/components/schemas/EnrichmentWorkerStatus' }
            deezer_rich:    { $ref: '#/components/schemas/EnrichmentWorkerStatus' }
            spotify_id:     { $ref: '#/components/schemas/EnrichmentWorkerStatus' }
            spotify_genres: { $ref: '#/components/schemas/EnrichmentWorkerStatus' }
            lastfm_id:      { $ref: '#/components/schemas/EnrichmentWorkerStatus' }
            lastfm_tags:    { $ref: '#/components/schemas/EnrichmentWorkerStatus' }
            lastfm_stats:   { $ref: '#/components/schemas/EnrichmentWorkerStatus' }

    # ── Library Browse ────────────────────────────────────────────────────────
...